Extracted source in 'utils' folder to a static lib
