+++ PixelBooster.pro
@@ -1,4 +1,5 @@
 TEMPLATE = subdirs
 
 SUBDIRS += \
-    View
+    View \ 
+    UtilsLib
