+++ UtilsLib/UtilsLib.pro
@@ -1,0 +1,15 @@
+#-------------------------------------------------
+#
+# Project created by QtCreator 2016-06-24T12:34:19
+#
+#-------------------------------------------------
+
+#QT       -= gui
+
+TARGET = UtilsLib
+TEMPLATE = lib
+CONFIG += staticlib
+
+SOURCES += pb_math.cpp
+
+HEADERS += pb_math.h
