--- View/utils/pb_math.h
@@ -1,33 +1,0 @@
-/***************************************************************************\
-*  Pixel::Booster, a simple pixel art image editor.                         *
-*  Copyright (C) 2015  Ricardo Bustamante de Queiroz (ricardo@busta.com.br) *
-*  Visit the Official Homepage: pixel.busta.com.br                          *
-*                                                                           *
-*  This program is free software: you can redistribute it and/or modify     *
-*  it under the terms of the GNU General Public License as published by     *
-*  the Free Software Foundation, either version 3 of the License, or        *
-*  (at your option) any later version.                                      *
-*                                                                           *
-*  This program is distributed in the hope that it will be useful,          *
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of           *
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the            *
-*  GNU General Public License for more details.                             *
-*                                                                           *
-*  You should have received a copy of the GNU General Public License        *
-*  along with this program.  If not, see <http://www.gnu.org/licenses/>.    *
-\***************************************************************************/
-
-#ifndef PB_MATH_H
-#define PB_MATH_H
-
-#include <QColor>
-#include <cmath>
-
-template <typename T>
-T clamp(const T &v, const T &lower, const T &upper) {
-  return (std::max)(lower, (std::min)(upper, v));
-}
-
-QColor ColorLerp(QColor &c1, QColor &c2, float t);
-
-#endif // PB_MATH_H
